[Bug] Setting a ttl on cached values in Streamlit (#21)
## Description
Fixes #20. Some cached values in Streamlit go stale. This includes the
db connection. After long periods of time if a user interacts with the
app it will attempt to reconnect but be stuck forever in a loop.

## Implementation
Setting a ttl. For the db connection set a 5 min timeout because that is
the same length as the serverless warehouse auto-shutdown interval.
For other data, set a longer timeout since these values don't get
updated frequently.

## Testing
Tested locally. Waited 5 min and tried interacting with the dashboard
again.