[Enhancement] Add the ability to look for a streamer (#26)
## Description
Closes #25 Adds the ability to search for a streamer

![image](https://github.com/user-attachments/assets/c60a4779-18f6-4ab9-bf03-bce11766a2f4)

The searchbox displays a list of streamers in the past week. Upon
picking a streamer the viewers for that streamer are obtained.

## Implementation
The list of unique streamers are precomputed via a materialized table
which is used to populate the searchbox. The searchbox is placed in
fragment which allows redrawing the graph every time a new streamer is
selected. The viewers are obtained by querying the silver_twitch_streams
table.

This PR also contains changes to increase the caching time to decrease
how often the sql warehouse is spun up.
## Testing
Tested locally